+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/favor/FavorAdder.java
@@ -23,7 +23,7 @@ public class FavorAdder implements ISlashRegistryInit {
         none.drop_runes = false;
         none.drop_lvl_rewards = false;
         none.drop_unique_gears = false;
-        none.drop_exp = true;
+        none.exp_multi = 0.25F;
         none.favor_drain_per_item = 0;
 
         none.locname = "Empty";
@@ -52,7 +52,7 @@ public class FavorAdder implements ISlashRegistryInit {
         FavorRank high = new FavorRank("high");
         high.min = 1000;
         high.rank = 3;
-        high.extra_item_favor_cost = 1.05F;
+        high.extra_item_favor_cost = 1.1F;
         high.extra_items_per_boss = 3;
         high.extra_items_per_chest = 2;
 
@@ -64,9 +64,10 @@ public class FavorAdder implements ISlashRegistryInit {
         FavorRank veryhigh = new FavorRank("very_high");
         veryhigh.min = 2000;
         veryhigh.rank = 4;
-        veryhigh.extra_item_favor_cost = 1.1F;
+        veryhigh.extra_item_favor_cost = 1.25F;
         veryhigh.extra_items_per_boss = 5;
         veryhigh.extra_items_per_chest = 3;
+        veryhigh.exp_multi = 1.1F;
 
         veryhigh.locname = "Very High";
         veryhigh.text_format = Formatting.BLUE.getName();
@@ -76,9 +77,10 @@ public class FavorAdder implements ISlashRegistryInit {
         FavorRank favored = new FavorRank("favored");
         favored.min = 5000;
         favored.rank = 5;
-        favored.extra_item_favor_cost = 1.5F;
-        favored.extra_items_per_boss = 10;
+        favored.extra_item_favor_cost = 1.75F;
+        favored.extra_items_per_boss = 7;
         favored.extra_items_per_chest = 5;
+        favored.exp_multi = 1.25F;
 
         favored.locname = "Favored";
         favored.text_format = Formatting.LIGHT_PURPLE.getName();