+++ src/main/java/com/robertx22/age_of_exile/saveclasses/player_skills/PlayerSkillData.java
@@ -43,6 +43,6 @@ public class PlayerSkillData {
     }
 
     public int getExpNeededToLevel() {
-        return (int) (LevelUtils.getExpRequiredForLevel(lvl + 1) * 0.25F);
+        return (LevelUtils.getExpNeededForSkillLevel(lvl + 1));
     }
 }